+++ plugin/src/main/kotlin/org/pcsoft/intellij/plugin/inno_setup/language/feature/editor/IsStructureAwareNavbar.kt
@@ -14,14 +14,7 @@ package org.pcsoft.intellij.plugin.inno_setup.language.feature.editor
 
 import com.intellij.ide.navigationToolbar.StructureAwareNavBarModelExtension
 import com.intellij.lang.Language
-import org.pcsoft.intellij.plugin.inno_setup.script.language.file_type.IsIcons
-import org.pcsoft.intellij.plugin.inno_setup.script.language.file_type.IsScriptFile
 import org.pcsoft.intellij.plugin.inno_setup.script.language.file_type.IsScriptLanguage
-import org.pcsoft.intellij.plugin.inno_setup.script.language.parser.section.displayName
-import org.pcsoft.intellij.plugin.inno_setup.script.language.parser.section.nameText
-import org.pcsoft.intellij.plugin.inno_setup.script.language.parser.section.psi.IsSectionBlock
-import org.pcsoft.intellij.plugin.inno_setup.script.language.parser.section.psi.IsSectionDirectiveEntry
-import org.pcsoft.intellij.plugin.inno_setup.script.language.parser.section.psi.IsSectionParameterEntry
 import javax.swing.Icon
 
 /**
@@ -37,23 +30,10 @@ class IsStructureAwareNavbar : StructureAwareNavBarModelExtension() {
     /**
      * Returns presentation metadata used by IntelliJ navigation UI.
      */
-    override fun getPresentableText(obj: Any?): String? {
-        return when (obj) {
-            is IsScriptFile -> obj.name
-            is IsSectionBlock -> obj.nameText
-            is IsSectionDirectiveEntry -> obj.keyText()
-            is IsSectionParameterEntry -> obj.displayName
-            else -> null
-        }
-    }
+    override fun getPresentableText(obj: Any?): String? = IsElementPresentation.textOf(obj)
 
     /**
      * Returns the icon shown for this element or file type.
      */
-    override fun getIcon(obj: Any?): Icon? = when (obj) {
-        is IsSectionBlock -> IsIcons.Section
-        is IsSectionDirectiveEntry -> IsIcons.ParameterEntry
-        is IsSectionParameterEntry -> IsIcons.ParameterEntry
-        else -> null
-    }
+    override fun getIcon(obj: Any?): Icon? = IsElementPresentation.iconOf(obj)
 }